fix: correct path to background image
This finally fixes the correct path to the background image, using a
relative URL. The fundamental cause was that launching this locally
hosts the site at the top-level of the domain, but the public site is
hosted at a sub-path of the domain (path = "framily-tree/").

To mitigate similar mistakes in the future, I've added a symlink to the
project directory named "mount-path." I can locally test how the site
will look when hosted at a sub-path by visiting
"http://127.0.0.1:8080/mount-path/" instead.

Fixes #5