feat: search for a particular brother
This adds a search box to the HTML, as well as functionality to search
for an arbitrary brother in the graph.

Search queries can be any substring of a brother's name,
case-insensitive. If a search succeeds, the graph zooms in and selects
the brother's node. If a search fails, the search text is highlighted in
red.

Test: manual