chore: drop node v10 support
No change to logic. This drops support for node v10 and sets node v12 as
the minimum supported version. This slightly speeds up presubmit testing
since that's one fewer node version to test. This also updates 2
dependencies which could not be updated before due to the need for node
v10 support.