refactor: remove dead code for error handling
This removes some dead code which does some error handling. These error
cases should be obsolete now that we have tests which parse the full
brothers.js file to ensure we can parse successfully.

This also adds `/* istanbul ignore next */` statements for some code
which is not testable: either because it is for error cases which are
not expected, or because it is for browser-only code (tests run in
NodeJS, so they won't ever be able to cover browser-only code).

Test: npm test
Test: npm start (verify that this runs OK in the browser)